requirements.txt
asgiref==3.7.2
Django==4.2.8
django-extensions==3.2.3
djangorestframework==3.14.0
numpy==1.26.2
pandas==2.1.4
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
